SQL script to detect damaged append-optimized tables
In one Greenplum 5X environment, it was found that an aoseg table's
relfilenode on primary was empty but its copy on the mirror was not
empty.  And the primary and mirror pair was in-sync.  Persistent
tables - gp_relation_node and gp_persistent_relation_node and the
filesystem on primary indicated that the AO table had one or more
segment files on disk.  Select on the AO table was failing with
"gp_relation_node has more entries than ..." error.

After copying the aoseg table's relfilenode from mirror into primary
data directory, the error was gone and select started working.

This SQL script is an attempt to detect such damaged AO tables by
comparing contents of aoseg and gp_relation_node for all
append-optimized tables in the system.  Knowing the extent of the
damage is the first step towards fixing it.

Note that before running the SQL script, it is assumed that gpcheckcat
has reported no catalog issues.